requirements.txt
langchain
bs4
chromadb
openai
tiktoken
sentence_transformers
unstructured
tabulate
pdf2image
faiss-cpu
